+++ extra/patterns/index.js
@@ -1,0 +1,4 @@
+import createDimondShape from './diamond.js';
+import  triangle from './triangle.js';
+
+export {createDimondShape,triangle};
