+++ extra/patterns/triangle.js
@@ -1,26 +1,12 @@
-k = 0;
-let s=" ";
-let flag = 1;
-// rows=4;
-rows = parseInt(process.argv[2]);
-// console.log(rows);
-for(i=1;i<=2*rows-1;i++){
-    if(i<rows-k)
-    process.stdout.write(" ");
-    else{
-        if(flag)
-            process.stdout.write("*");
-        else
-            process.stdout.write(" ");
-        flag=1-flag;
-        }
-    if(i==rows+k)
-    {
-        k++;
-        console.log("\n");
-        if(i==2*rows-1)
-        break;
-        i=0;
-        flag = 1;
-    }
-}
+export default function triangle(n) { 
+    for (var i = 0; i < n; i++) { 
+      var str = ''; 
+      for (var j = 1; j < n-i; j++) { 
+        str = str + ' '; 
+      } 
+      for (var k = 1; k <= (2*i+1); k++) { 
+        str = str + '*'; 
+      } 
+      console.log(str); 
+    } 
+  }
