+++ extra/utils/validation.js
@@ -1,35 +1,5 @@
-const user =
-    [
-        {
-            traineeEmail: 'trainee1@successive.tech',
-            reviewerEmail: 'reviewer1@successive.tech',
-        },
-        {
-            traineeEmail: 'trainee2@successive.tech',
-            reviewerEmail: 'reviewer2@successive.tech',
-        },
-        {
-            traineeEmail: 'trainee3uccessive.tech',
-            reviewerEmail: 'reviewer3@successive.tech',
-        },
-        {
-            traineeEmail: 'trainee4@successive.tech',
-            reviewerEmail: 'reviewer4uccessive.tech',
-        },
-        {
-            traineeEmail: 'trainee5@gmail.tech',
-            reviewerEmail: 'reviewer5@successive.tech',
-        }
-    ]
-
-function validateEmail(Email){
-RegExp="^[a-zA-Z0-9+_.-]+@successive.tech+$";
-let str=Boolean( Email.match(RegExp));
-return str;
-}        
-
-
-function validateUsers(user){
+import validateEmail from './helper';
+export default function validateUsers(user){
 let valid = 0;
 let invalid = 0;
 
@@ -49,4 +19,4 @@ console.log("Invalid Users Count : ",invalid);
 
 }
 
-validateUsers(user);
+// validateUsers(user);
